+++ python/paddle/fluid/tests/unittests/test_dist_mnist.py
@@ -26,56 +26,6 @@ class TestDistMnist2x2(TestDistBase):
         self.check_with_place("dist_mnist.py", delta=1e-5)
 
 
-class TestDistMnistNCCL2(TestDistBase):
-    def _setup_config(self):
-        self._sync_mode = True
-        self._use_reduce = False
-        self._use_reader_alloc = False
-        self._nccl2_mode = True
-
-    def test_dist_train(self):
-        import paddle.fluid as fluid
-        if fluid.core.is_compiled_with_cuda():
-            self.check_with_place("dist_mnist.py", delta=1e-5)
-
-
-class TestDistMnistNCCL2MultiNCCLComm(TestDistBase):
-    def _setup_config(self):
-        self._sync_mode = True
-        self._use_reduce = False
-        self._use_reader_alloc = False
-        self._nccl2_mode = True
-        self._nccl_comm_num = 3
-
-    def test_dist_train(self):
-        import paddle.fluid as fluid
-        if fluid.core.is_compiled_with_cuda():
-            self.check_with_place("dist_mnist.py", delta=1e-5)
-
-
-class TestDistMnistNCCL2DGC(TestDistBase):
-    def _setup_config(self):
-        self._sync_mode = True
-        self._use_reduce = False
-        self._use_reader_alloc = False
-        self._nccl2_mode = True
-        self._use_dgc = True
-
-    def test_dist_train(self):
-        import paddle.fluid as fluid
-        if fluid.core.is_compiled_with_cuda():
-            self.check_with_place("dist_mnist.py", delta=1e-5)
-
-
-class TestDistMnist2x2Lars(TestDistBase):
-    def _setup_config(self):
-        self._sync_mode = True
-        self._use_reduce = False
-
-    def test_se_resnext(self):
-        self.check_with_place("dist_mnist_lars.py", delta=1e-5)
-
-
 class TestDistMnist2x2WithMemopt(TestDistBase):
     def _setup_config(self):
         self._sync_mode = True
